Fix broken services in HA 2024.2.1
In Home Assistant `2024.2.1` this service is always erroring with:

```
Failed to call service opensprinkler/run. 'list' object has no attribute 'get'
```

The stacktrace for this error is below:

```
2024-02-10 12:46:23.684 ERROR (MainThread) [homeassistant.components.websocket_api.http.connection] [<REDACTED>] 'list' object has no attribute 'get'
Traceback (most recent call last):
  File "/usr/src/homeassistant/homeassistant/components/websocket_api/commands.py", line 240, in handle_call_service
    response = await hass.services.async_call(
               ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/src/homeassistant/homeassistant/core.py", line 2279, in async_call
    response_data = await coro
                    ^^^^^^^^^^
  File "/usr/src/homeassistant/homeassistant/core.py", line 2316, in _execute_service
    return await target(service_call)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/config/custom_components/opensprinkler/__init__.py", line 108, in _async_send_run_command
    await hass.helpers.service.entity_service_call(
  File "/usr/src/homeassistant/homeassistant/helpers/service.py", line 843, in entity_service_call
    entity_candidates = _get_permissible_entity_candidates(
                        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/usr/src/homeassistant/homeassistant/helpers/service.py", line 796, in _get_permissible_entity_candidates
    and (entity := entities.get(single_entity)) is not None
                   ^^^^^^^^^^^^
AttributeError: 'list' object has no attribute 'get'
```

After some investigation I found that the function signature for
`entity_service_call` was changed in
https://github.com/home-assistant/core/pull/106759 from
`Iterable[EntityPlatform]` to `dict[str, Entity]`.

It's not clear to me if this is some internal API that cannot be relied
upon by custom components as it does not seem this was a widely
communicated API change. But in any case it seems this change fixes the
problem for me and this new function was copied from a similar change in
that PR.